Create 2021_day6.py
Initially I went about updating a data frame for this problem to practice pandas, and also see if it could be done. It actually worked quite well and was quick for part 1 but couldn't handle 256 days of Part 2. I redid the solution with a map using dictionaries which was the much more efficient approach.