Create 2021_day9.py
I learned a lot but needed a ton of help on this one in the form of perusing the solutions thread on reddit. The first part was easy enough but I had to learn how to optimize for the second part. My main contribution to the second part was redoing the solutions to use default dicts and numpy, to speed up the operation and make them look neater.